requirements.txt
torch
transformers
openai-whisper
TTS
sounddevice
soundfile
numpy
requests